requirements.txt
django>=2.1.5
djangorestframework>=3.9.1
requests>=2.21.0
lxml>=4.3.0
